requirements.txt
python-telegram-bot>=20.0
yandex-music>=3.0.0
pytest>=7.0.0
